# lapack/testlapack: add benchmark for Dlangb