--- a/lapack/gonum/bench_test.go
+++ b/lapack/gonum/bench_test.go
@@ -11,4 +11,5 @@ import (
 )
 
 func BenchmarkDgeev(b *testing.B)  { testlapack.DgeevBenchmark(b, impl) }
+func BenchmarkDlangb(b *testing.B) { testlapack.DlangbBenchmark(b, impl) }
 func BenchmarkDlantb(b *testing.B) { testlapack.DlantbBenchmark(b, impl) }