Comparison between Dlangb implemented with work slice and without work slice:
```
» benchstat dlangb-bench-with-work.txt dlangb-bench-without-work.txt 
name                              old time/op  new time/op   delta
Dlangb/MaxColSum_N=1000_K=0-4     5.57µs ± 5%   9.80µs ± 1%  +75.91%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=1000_K=1-4     8.08µs ± 0%  13.15µs ± 2%  +62.71%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=1000_K=2-4     10.4µs ± 0%   16.3µs ± 0%  +56.56%  (p=0.029 n=4+4)
Dlangb/MaxColSum_N=1000_K=30-4    71.6µs ± 0%   44.4µs ± 2%  -37.95%  (p=0.016 n=4+5)
Dlangb/MaxColSum_N=10000_K=0-4    48.5µs ± 3%   79.8µs ± 0%  +64.74%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=1-4    71.5µs ± 3%  117.3µs ± 2%  +64.17%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=2-4    92.8µs ± 0%  153.3µs ± 0%  +65.17%  (p=0.029 n=4+4)
Dlangb/MaxColSum_N=10000_K=100-4  2.52ms ± 1%   1.92ms ± 0%  -23.82%  (p=0.008 n=5+5)
```
The result is inconclusive, for small bandwidths the work-slice version is faster, for larger bandwidths the version without the work slice (current master) is faster. Maybe I should add more bandwidths to see where the tipping point is.

Closes #1624 .

Please take a look.


<!--
Checklist:

- API changes have been discussed
- code is goformated correctly (goimports)
- packages with generated code have had code generation run
- tests pass locally
- linked to relevant issues

Please make sure your commit message summary line and pull request
title match the Go convention; a one-line summary of the change,
prefixed by the primary affected package that should complete the
sentence, "This change modifies Gonum to _____."
-->