@vladimir-ch 2021-05-26
I've added more bandwidth, so for MaxCol the results are like this now:
```
name                              old time/op  new time/op   delta
Dlangb/MaxColSum_N=1000_K=0-4     4.60µs ± 0%   8.53µs ± 5%  +85.59%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=1000_K=1-4     6.65µs ± 0%  12.34µs ± 2%  +85.71%  (p=0.016 n=4+5)
Dlangb/MaxColSum_N=1000_K=2-4     8.87µs ± 2%  16.96µs ±12%  +91.15%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=1000_K=5-4     15.5µs ± 2%   16.8µs ± 4%   +8.44%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=1000_K=8-4     21.6µs ± 1%   16.8µs ± 5%  -22.31%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=1000_K=10-4    25.4µs ± 0%   24.4µs ± 4%     ~     (p=0.095 n=5+5)
Dlangb/MaxColSum_N=1000_K=20-4    52.2µs ± 0%   26.8µs ± 1%  -48.64%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=1000_K=30-4    66.7µs ± 0%   41.3µs ± 2%  -38.12%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=0-4    45.7µs ± 0%   85.1µs ± 3%  +86.43%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=1-4    67.1µs ± 0%  122.9µs ± 0%  +83.04%  (p=0.029 n=4+4)
Dlangb/MaxColSum_N=10000_K=2-4    89.1µs ± 0%  161.8µs ± 1%  +81.64%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=5-4     152µs ± 0%    168µs ± 2%  +10.77%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=8-4     216µs ± 1%    167µs ± 2%  -22.91%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=10-4    257µs ± 1%    243µs ± 2%   -5.72%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=30-4    764µs ± 0%    670µs ± 4%  -12.35%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=60-4   1.58ms ± 2%   1.23ms ± 2%  -22.44%  (p=0.008 n=5+5)
Dlangb/MaxColSum_N=10000_K=100-4  2.26ms ± 1%   1.31ms ± 4%  -42.05%  (p=0.008 n=5+5)
```
There is clearly an overhead for calling into `f64` that pays off only for larger bandwidth. I would leave it as is, I don't think Dlangb is called in any inner loop.